+++ transform.h
@@ -48,19 +48,8 @@ static inline CcTransform cc_transform_identity(void)
     return t;
 }
 
-static inline CcTransform cc_transform_inverse(CcTransform t)
-{
-    double det = (t.m[0] * t.m[3] - t.m[1] * t.m[2]);
-    double s = 1.0 / det;
-
-    CcTransform inv = {
-        { s * t.m[3], s * -t.m[1], s * -t.m[2], s * t.m[0] }
-    };
-
-    return inv;
-}
-
-static inline CcTransform cc_transform_rotate(double angle)
+static inline
+CcTransform cc_transform_rotate(double angle)
 {
     double c = cos(angle);
     double s = sin(angle);
@@ -81,34 +70,6 @@ CcTransform cc_transform_scale(double w, double h)
 }
 
 static inline
-CcTransform cc_transform_skew(double x_angle, double y_angle)
-{
-    double c_x = cos(x_angle);
-    double s_x = sin(x_angle);
-
-    double c_y = cos(y_angle);
-    double s_y = sin(y_angle);
-
-    CcTransform t = {
-        {  c_x, s_x, -s_y, c_y }
-    };
-    return t;
-}
-
-static inline
-CcTransform cc_transform_concat(CcTransform a, CcTransform b)
-{
-    CcTransform t = {
-        { a.m[0] * b.m[0] + a.m[1] * b.m[2],
-          a.m[0] * b.m[1] + a.m[1] * b.m[3],
-          a.m[2] * b.m[0] + a.m[3] * b.m[2],
-          a.m[2] * b.m[1] + a.m[3] * b.m[3],
-        }
-    };
-    return t;
-}
-
-static inline
 Vec2 cc_transform_apply(CcTransform t, Vec2 v)
 {
     Vec2 r;
@@ -116,6 +77,10 @@ Vec2 cc_transform_apply(CcTransform t, Vec2 v)
     r.y = t.m[2] * v.x + t.m[3] * v.y;
     return r;
 }
+
+CcTransform cc_transform_inverse(CcTransform t);
+CcTransform cc_transform_skew(double x_angle, double y_angle);
+CcTransform cc_transform_concat(CcTransform a, CcTransform b);
 
 CcBitmap* cc_bitmap_transform(const CcBitmap* src, CcBitmap* dst, CcTransform A, uint32_t bg_color);
